requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.58.1,<2.0.0
numpy==1.26.4
python-dotenv==1.0.0
pydantic==2.12.5